requirements.txt
pandas
openpyxl
PyMuPDF
pdfplumber
openai
python-dotenv
tqdm
